// src/Model/Movie.ts
import Member from './Member';

export default interface Movie {
	actors: Member[];
	adult: boolean;
	backdrop_path: string;
	belongs_to_collection: any;
	budget: number;
	directors: Member[];
	genres: any[];
	homepage: string;
	id: number;
	imdb_id: string;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	production_companies: any[];
	production_countries: any[];
	release_date: string;
	revenue: number;
	runtime: number;
	spoken_languages: any[];
	status: string;
	tagline: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}